finishing pay method

+++ src/app/pay-method/paymethod-item/paymethod-item.component.ts
@@ -3,6 +3,7 @@ import { PaymethodService } from '../../services/paymethod/paymethod.service';
 import {ToastService} from '../../services/toast/toast.service';
 import {MdDialog, MdDialogRef} from '@angular/material';
 import {PaymethodEditComponent } from '../paymethod-edit/paymethod-edit.component';
+import { ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-paymethod-item',
@@ -28,29 +29,39 @@ export class PaymethodItemComponent implements OnInit {
   }
 
   removePaymethod = (elem) => {
-    this.payMethodService.delete(elem.id)
-      .subscribe(
-          data => {
-              this.updateView.emit();
-              this.toastService.success({message: 'Pay Method Deleted!'});
-          },
-          (err) => {
-              this.toastService.error({message: 'An error has occur!'});
-          }
-      )
+    var message  = 'The pay method "' + this.payMethod.name + '" will be deleted!';
+
+    let dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '300px'
+    });
+    dialogRef.componentInstance.message = message; //Passing data to the Dialog, this is received as 'payMethod'
+    dialogRef.afterClosed().subscribe(response => {
+      if(response){
+        this.payMethodService.delete(elem.id)
+          .subscribe(
+              data => {
+                  this.updateView.emit();
+                  this.toastService.success({message: 'Pay Method Deleted!'});
+              },
+              (err) => {
+                  this.toastService.error({message: 'An error has occur!'});
+              }
+          )
+      }
+    })
   }
 
   editPaymethod(payMethod){
-    let dialogRef = this.dialog.open(PaymethodEditComponent);
+    let dialogRef = this.dialog.open(PaymethodEditComponent, {
+      width: '300px'
+    });
 
-    
     dialogRef.componentInstance.payMethod = payMethod; //Passing data to the Dialog, this is received as 'payMethod'
     dialogRef.afterClosed().subscribe(data => {
       if(data){
         this.updateView.emit();
       }
     });
-    
   }
 
 }
